models refactor. clean initial migration
