+++ indieauth/models/indieauth_base.py
@@ -1,0 +1,2 @@
+class IndieAuthBase():
+    pass
